wip: Add test for one spare and refactor
Adds pending test for one spare rolled. Adds wip refactoring bowling
game so that game.score calculates score and rolls adds pins knocked
to array of rolls.

This concludes my 30 minute kata practice. I was able to perform the
first two tests in the exact process taught in the slides. Then, I
completed the thrid test following the slides carefuly while studying
the steps and reasons behind the steps. After completing the third test
and committing, I reverted the commit to practice the third test again.
I was able to make it to slide 34--part way through the process of
refactoring with the third test in a pending state.

Tomorrow, I will revert this commit to start at the beginning of the
third test to get more practice at this section of the kata.