Bowling game passes thrid test
- Adds test for score one spare
- Refactors game.rolls and game.score
- Modifies BowlingGame to pass test

I was able to practice the third test twice today and I believe I have
a firm understanding of the key concepts of this phase.

There are three concepts practiced in this phase. One being the practice
of skipping a failing test while refactoring code in preparation to pass the
failing test. Observing this practice allows one to incrementally modify
the code and quickly verify that a change is not a regression. Another,
good practice is refactoring in the smallest possible increments that
are able to pass all tests. Finally, carefully regard setting flags as
a potential smell of poor design.

Tommorrow, I will perform the kata from the beginning before proceeding
to work on mastering the fourth test phase.