Bowling game passes strike test. Needs refactor.
Adds test for strike. BowlingGame is modified by needs additional
refactoring. strikeBonus function should be factored from score method.

I performed the first three tests/phases of the bowling game kata
perfectly from memory for today's practice. Then, I used the last 8 or
so minutes to begin studying the fourth test/phase. I nearly completed
it in the time given.

Tomorrow, I will revert this last commit to reset the game to the
start of test/phase 4 and focus on the fourth segment of this kata.