Bowling kata passing test 2 (spare)
Today I focused on studying the fourth test. After going through it twice,
I completed the fifth test which passed immediately. Then, I reset
the kata and started from the beginning progressing as quickly as
possible with the remaining time.

I was able to complete the second test, refactor, and write a pending
test for the "one spare" case.

Tommorrow I will start from this point and repeat the kata until my
25 minute time block for kata practice concludes.