Score-keeper renders score and scores one and two
This commit adds a nearly complete score keeper. Call `createScoreKeeper`
to create an instance of a scoreKeeper. The scoreKeeper object has
a printScore method which returns a string representation of the score
with exactly 3 digits for the score for each team padded with zeroes
like "009:002". The scoreKeeper has methods to add one or two points
to a team's score.

TODO:
Finish refactoring and add methods to score 3 points for a team.