Complete score keeper kata
I didn't quite follow the specifications because I wished to use
a functional-ish style. All the score methods take a score as input and
return a new score as output. This effectively removes state to the
thing invoking the methods rather than tracking the state in an object.

This makes all methods in scoreKeeper idempotent. Not sure if that
matters. However, it was a different way to approach the problem.

The one difference that seemed to make writing the tests and code easier
was completely separating the function for rendering the score as a
string from the ScoreKeeper.