Complete score-keeper kata
Complete the score keeper kata. This time, I returned to using a factory
method to create the scoreKeeper object. During the process of factoring
out duplicate code for the addition of each scoring method, I have
a tendancy to normalize the code so far that I fear it may be hard to
comprehend the code. I believe I need to strike a better balance between
a procedural vs highly normalized functional approach. This seems to be
the pitfall of the "Single Responsibility Principle" (SRP). It is
possible to define a responsiblity so narrowly that it is no different
than a single instruction. Perhaps reminding myself that SRP is about
"one reason to change", according to Bob Martin. Hmmmm...