Score keeper kata wip
Score keeper exports printScore method and addOneForTeamA method.

Invoking printScore returns the score rendered as a string. Each score
is three characters in length with leading zeros as needed. (e.g '002:010')

Calling addOneForTeamA([0, 1]) returns [1, 1]

This version of the kata is a procedural/functional style approach
instead of using the object oriented approach to the kata.