Complete score-keeper kata
This complete a functional/procedural style for solving the score
keeper kata. In order to have no side effects with the functions, the
controller which calls the scoreKeeper methods is responsible for
storing the state of score rather than the prescribed approach in the
prompt which specifies that each method updates the state of the score
as a side effect. The benefit of the stateless approach is that the
manner by which state is preserved is completely decoupled from the
logic for scoring and rendering the score.