GOL: passing test for dead neighbors (green)
This marks the end of my 25 minute practice. I'm creating a commit for
each step of the red/green/(refactor) cycle. There is some refactoring
needed before adding the next test.

There is something I don't care for about this kata. It seems there is
an assumption that a Grid class will be created. However, I just don't
see the need for it. A functional approach seems to work just fine when
testing from the "Outside-in". My reading on this term suggest that
an "Outside-in" approach will likely involve mocking lower level
functions/classes before developing them with their own suite of tests.
However, GOL doesn't seem to required this mocking unless the Grid class
is used for implementing a solution. The solutions I've found on Github
which use a Grid class (as sometimes several others) seem far too over
complicated. Sure, they might be slightly easier to extend, but the
solutions I've developed thus far would not be hard to refactor to use
a Grid class and then extend. I feel that someone with no previous
knowledge could affect such modifications more easily when presented
the functional code than try to understand the complex web of classes that
so many seem to prefer before implementing modifications.