GOL: Add pending test and skip prev test (green)
This commit adds a new test which will be useful but requires to
many changes. I believe some intermediate tests should be added first.
This test also revealed that the preceding test suffers the same problem.
If this were an actual project for a client, I would probably write the
code to pass these tests for the sake of time management since the tests
are validiting correct behavior. However, for the purpose of a kata,
I want to improve at identifying a progressive sequence of tests that
make each red/green/refactor iteration as simple as possible while
making the source code more general and the tests more specific.