SP: add test 2: [1, 2] [red]
This is the end of my 25 minute practice block. I intentionally did not
progress far today. I made many commits that I normally would have
lumped into a single commit. My reason for proceeding in this way is to
document the steps of "red", "green", and maybed "refactor" though the
process of writing the first test and the code to pass it. There are a
lot of red/green cycles involved in writing setup code for the test
runner which imports libraries and the module exporting the code to
test.